requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
Authlib==0.13
black==19.10b0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
flask-marshmallow==0.10.1
Flask-OAuthlib==0.9.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
oauthlib==2.1.0
pathspec==0.6.0
pycparser==2.19
PyJWT==1.7.1
pymodm==0.4.2
pymongo==3.10.0
pyOpenSSL==19.0.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
regex==2019.12.19
requests==2.21.0
requests-oauthlib==1.3.0
six==1.13.0
SQLAlchemy==1.3.12
toml==0.10.0
typed-ast==1.4.0
urllib3==1.24.3
Werkzeug==0.16.0
